Create NaughtsAndCrossesAi.js
Link to source code in p5.js: https://editor.p5js.org/Adsa2/sketches/PzzxPIh7f

This is a program I wrote in my first year at university as I was interested in the idea of move generation algorithms (i.e chess engines). I wrote this to see if I could create an unbeatable noughts and crosses Ai from my understanding of recursive algorithms. Since this was more of a proof of concept, the code is not very optimised and there are a number of improvements that could be made:

- Improve the gameOver function (function to return who has won in a given game instance). This is not a very eloquent solution and is also not efficient. An easy improvement would be to return null automatically if less than 5 moves have been made (since it would be impossible to have won)

- Improve the move selection when presented with identical game values. At the moment, the choice between two winning moves is random. However, the move that results in the shortest forced win should be selected (has the amusing effect of stalling out the game but ensuring it can always still win). Furthermore, the choice between two drawing moves should favour the move that encounters the most winning game states during the min max algorithm.

-  Reduce the redundant searches made. A common improvement to minmax algorithms is Alpha Beta pruning, where the total searches is reduced

- remove continual evaluation of the same position. Since the game is extremely simple, there is no max depth limit so the whole game tree is computed on move 1. This could just be pre-processed and stored, and then the computer could look-up the best move in that position for each of its goes. This approach is used in connect 4 and also simple chess endgames (tablebase)