Create slit_fitting
This program was created to aid in my lab work as part of my physics degree, however it was well beyond the expected level of analysis needed for this particular lab experiment. The experiment involved measuring the diffraction pattern produced by a variety of slit groups, with varying slit width, slit separation and number of slits. Through analysis of the intensity-angle relationship, the diffraction pattern was fitted to a curve of equations found theoretically from the application of the Huygens–Fresnel principle. From this, estimations of the parameters mentioned could be found. 
The program also saves the plots in the folder called plots, which is also attached. Optically, it is clear that the fitted functions agree well with the measured data, especially for n (number of slits) >1.
One issue with this program is that the curve fitting approach can sometimes be sensitive to the choice of starting parameters. My solution was to manually set the starting parameters in these cases. However, a more general approach of estimating the starting parameters accurately would be preferred. Despite this, I am still happy with this approach since it worked well for our limited data set.